pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arch>=8.0.0",
    "enappsys==0.1.0",
    "fastparquet>=2025.12.0",
    "hydra-core>=1.3.2",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "numpy==2.4.1",
    "pandas==2.3.3",
    "python-dotenv==1.2.1",
    "requests==2.32.5",
    "scikit-learn>=1.8.0",
    "sklearn-preprocessing>=0.1.0",
    "statsmodels>=0.14.6",
    "torch>=2.10.0",
]
